+++ tests/TestCase/TestSuite/TestCaseTest.php
@@ -27,10 +27,10 @@ class TestCaseTest extends TestCase
      */
     public function testAssertImageFileEquals()
     {
-        $original = Configure::readOrFail(THUMBER . '.comparingDir') . 'resize_w200_h200.jpg';
+        $original = Configure::readOrFail('Thumber.comparingDir') . 'resize_w200_h200.jpg';
         $copy = tempnam(TMP, $original);
         copy($original, $copy);
-        $this->assertImageFileEquals(Configure::readOrFail(THUMBER . '.comparingDir') . 'resize_w200_h200.jpg', $copy);
+        $this->assertImageFileEquals(Configure::readOrFail('Thumber.comparingDir') . 'resize_w200_h200.jpg', $copy);
         $this->assertImageFileEquals('resize_w200_h200.jpg', $copy);
     }