+++ tests/test_app/config/routes.php
@@ -23,5 +23,5 @@
 use Cake\Routing\Router;
 
 Router::scope('/', function ($routes) {
-    $routes->loadPlugin(THUMBER);
+    $routes->loadPlugin('Thumber');
 });